--- models/schema.sql
@@ -1,19 +1,0 @@
-DROP DATABASE IF EXISTS volunteer_db;
-CREATE DATABASE volunteer_db;
-USE volunteer_db;   
-
-<<<<<<< HEAD
-CREATE TABLE volunteers (
-    id INT NOT NULL AUTO_INCREMENT,
-    FMID INTEGER(1100000) NOT NULL,
-    MarketName VARCHAR(200) NOT NULL,
-    Website VARCHAR(500) NOT NULL,
-    Street VARCHAR(100) NOT NULL,
-    City VARCHAR(50) NOT NULL,
-    County VARCHAR(50) NOT NULL,
-    State VARCHAR(50) NOT NULL,
-    Zip Code VARCHAR(50) NOT NULL,
-    PRIMARY KEY(id)
-);
-=======
->>>>>>> 422ee294edb6a3f33a2271281444ee642909be16
