+++ package.json
@@ -14,6 +14,8 @@
     "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "express-handlebars": "^3.0.0",
+    "inquirer": "^6.2.0",
+    "mysql": "^2.16.0",
     "mysql2": "^1.6.1",
     "sequelize": "^4.39.1"
   },
