+++ schema.sql
@@ -1,0 +1,4 @@
+DROP DATABASE IF EXISTS volunteer_db;
+CREATE DATABASE volunteer_db;
+USE volunteer_db;   
+
